trim code to match protocol; move all to subdir
