--- common_grants/__init__.py
@@ -1,64 +1,0 @@
-"""
-CommonGrants Python SDK
-
-A Python implementation of the CommonGrants protocol for grant data exchange.
-"""
-
-from .schemas.fields import (
-    # Field types
-    CustomField,
-    CustomFieldType,
-    DecimalString,
-    Event,
-    ISODate,
-    ISOTime,
-    Money,
-    SystemMetadata,
-    UTCDateTime,
-    Url,
-)
-from .schemas.models import (
-    # Models
-    ApplicationBase,
-    ApplicationStatus,
-    ApplicationStatusOptions,
-    AwardBase,
-    AwardStatus,
-    AwardStatusOptions,
-    Contact,
-    OppFunding,
-    OppStatus,
-    OppStatusOptions,
-    OppTimeline,
-    OpportunityBase,
-)
-
-__version__ = "0.1.0"
-
-__all__ = [
-    # Field types
-    "CustomField",
-    "CustomFieldType",
-    "DecimalString",
-    "Event",
-    "ISODate",
-    "ISOTime",
-    "Money",
-    "SystemMetadata",
-    "UTCDateTime",
-    "Url",
-    
-    # Models
-    "ApplicationBase",
-    "ApplicationStatus",
-    "ApplicationStatusOptions",
-    "AwardBase",
-    "AwardStatus",
-    "AwardStatusOptions",
-    "Contact",
-    "OppFunding",
-    "OppStatus",
-    "OppStatusOptions",
-    "OppTimeline",
-    "OpportunityBase",
-]
