--- common_grants/schemas/models/__init__.py
@@ -1,26 +1,0 @@
-"""Models for the CommonGrants API."""
-
-from .app_base import ApplicationBase
-from .app_status import ApplicationStatus, ApplicationStatusOptions
-from .award_base import AwardBase
-from .award_status import AwardStatus, AwardStatusOptions
-from .contact import Contact
-from .opp_base import OpportunityBase
-from .opp_funding import OppFunding
-from .opp_status import OppStatus, OppStatusOptions
-from .opp_timeline import OppTimeline
-
-__all__ = [
-    "ApplicationBase",
-    "ApplicationStatus",
-    "ApplicationStatusOptions",
-    "AwardBase",
-    "AwardStatus",
-    "AwardStatusOptions",
-    "Contact",
-    "OpportunityBase",
-    "OppFunding",
-    "OppStatus",
-    "OppStatusOptions",
-    "OppTimeline",
-]
