+++ lib/python-sdk/common_grants/__init__.py
@@ -1,0 +1,12 @@
+"""
+CommonGrants Python SDK
+
+A Python implementation of the CommonGrants protocol.
+"""
+
+__version__ = "0.1.0"
+
+from . import schemas
+
+__all__ = ["schemas"]
+
